% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sources}
\alias{data_sources}
\title{Data Sources: Descriptions, Download Links, etc.}
\format{A data frame with YY variables
\describe{
  \item{title}{Name of the dataset.}
  \item{abbrev}{Abbreviation used for file accessing for each dataset. Data with multiple versions (e.g. BCR v10.0) are designated using data_vXX (e.g., BCR_v10).}
  \item{description}{A brief description of the data.}
  \item{file_type}{Type of file(s)}
  \item{data_url}{}
  \item{license}{License name (DNE=does not exist)}
  \item{data_usage}{Categorical. Team-imposed categories to describe the license level (e.g, CC0==open) }
  \item{refresh_rate_mos}{Frequency of data updating, unit=months)}
  \item{primary_curator}{Institution hosting or curating the data. Otherwise the primary contact organization to learn more about the data.}
  \item{context}{Categorical. Team-imposed categories to describe the type of data.(administrative_region=top-down decision making influencing on-ground decisions with authority; advisory_region=top-down advisory councils, not necessarily on-ground actions influenced by decisions; managed_lands=actively managed or conserved property; integrated_data=heterogeneous data sources, multiple sources).}
  \item{spat_extent}{Spatial extent of the information.}
  \item{spat_resolution}{Spatial resolution of the information}
  \item{temporal_extent}{Temporal extent of the information}
  \item{temporal_resolution}{Temporal resolution of the information}
  \item{in_padus}{Logical. Is the data in PAD-US (USGS Protected Areas Database)?}
  \item{in_sfr}{Logical. Is the data in SFR (USGS Spatial Features Registry)}
  \item{notes}{Team notes and relevant information}
  \item{high_level_url}{URL to learn more about the dataset and/or organization(s).}
}}
\source{
\url{https://github.com/TrashBirdEcology/USAvian/blob/master/data/data_sources.csv}
}
\usage{
data_sources
}
\description{
A data frame containing the data sources used in the pilot tool, USAvian
}
\keyword{datasets}
