% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data from URL to local destination.}
\usage{
download_data(dir = NULL, url = NULL, fn = NULL, abbrevs = NULL)
}
\arguments{
\item{dir}{(optional) Directory for storing the downloaded data. Defaults to `data/data-raw`}

\item{url}{(optional) URL(s) character, single or vector, for data source locations.}

\item{fn}{(optional) Filename(s) corresponding to the name local filename of the downloaded data (do not include the extension).}

\item{abbrevs}{(Preferred) Abbreviations of select datasets (see col "abbrev" `data("data_sources")`)). Specifying this parameter is the preferred, and easier, method for downloading relevant data sources.}
}
\value{
NA. Files downloaded to local directory (as defined by `dir` parameter.)
}
\description{
Function provides an option to specify the data you want to download using URLs and desired filenames, or you can download files from our [Data Source List](https://github.com/TrashBirdEcology/USAvian/blob/master/data/data_sources.csv) and use the abbreviations to quickly download all desired data to file.
}
\examples{
\\donotrun{
# Download the persistent Bird Conservation Regions (BCR) compressed file.
download_data(abbrev="bcr_persistent")
}
}
